adding utf8_split_off

+++ src/syntax/mod.rs
@@ -100,7 +100,7 @@ impl Lexer {
     #[inline]
     pub fn render_modal_if_exist(&mut self, row: u16, col: u16, gs: &mut GlobalState) {
         if self.modal_rect.is_none() {
-            self.modal_rect = self.modal.as_mut().and_then(|modal| modal.render_at(col as u16, row as u16, gs));
+            self.modal_rect = self.modal.as_mut().and_then(|modal| modal.render_at(col, row, gs));
         };
     }
